backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-ares-backend"
version = "0.1.0"
description = "Project Ares backend scaffold"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.3",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]
